// src/shared/types.ts
export type UserType = {
    _id: string; // Assuming the ID is a string, change it to the appropriate type if needed
    id: number; // Assuming the ID is a number
    first_name: string;
    last_name: string;
    email: string;
    gender: string;
    avatar: string;
    domain: string;
    available: boolean;
}